+++ config/query-builder.php
@@ -1,4 +1,7 @@
 <?php
+/**
+ * @see https://github.com/spatie/laravel-query-builder
+ */
 
 return [